backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
ultralytics
Pillow
numpy
opencv-python-headless
